Cheesy hack for values not found in lookup options.
Ideally instead of this hack, this function and the validation before
it would be modified. Any time we assume that a value exists within
lookup options, I would think we would check that in the
pre-validation. Alternatively set an empty or NULL value.

This holds true for all of these lookups- for items and models too.