UI adjustments for model template
Rename "Model Detail" page to "Master Model Detail"
The "Derivative Models" data table should list the available derivative models
Clicking on a row in this data table should bring the user to a page for that derivative. For now we can just have one data table that list all the files we have for that derivative?
Change "Model Files" data table heading to "Geometry Files"